requirements.txt
openbb-agents
openbb
python-dotenv